added event handlers for textfields of new player

+++ app/src/main/java/com/explore/r11/domain/model/NewPlayer.kt
@@ -4,5 +4,5 @@ data class NewPlayer (
     var name:String,
     var team:String,
     var type:String,
-    var salary:Int
+    var salary:Double
 )
